Add dex manifests
Updated dex compilation system. Now each group of dexes has a manifest.json file that corresponds to it. These files map each dex type (species, item, etc.) to their location relative to the manifest.

All of these manifests are then registered in the masterManifest.json in the root.

The compile_masterdexes script was updated to reflect this.